agent/requirements.txt
livekit-agents
livekit-plugins-elevenlabs
livekit-plugins-langchain
langchain-mistralai
python-dotenv